+++ test/plots/stargazers-hourly.js
